[Kernel][Cleanup] Move table features related classes into a separate namespace
