+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/TableImpl.java
@@ -30,6 +30,7 @@ import io.delta.kernel.internal.fs.Path;
 import io.delta.kernel.internal.metrics.SnapshotQueryContext;
 import io.delta.kernel.internal.metrics.SnapshotReportImpl;
 import io.delta.kernel.internal.snapshot.SnapshotManager;
+import io.delta.kernel.internal.tablefeatures.TableFeatures;
 import io.delta.kernel.internal.util.Clock;
 import io.delta.kernel.metrics.SnapshotReport;
 import io.delta.kernel.types.StructField;
