+++ kernel/kernel-api/src/main/java/io/delta/kernel/internal/TransactionImpl.java
@@ -37,6 +37,7 @@ import io.delta.kernel.internal.metrics.TransactionReportImpl;
 import io.delta.kernel.internal.replay.ConflictChecker;
 import io.delta.kernel.internal.replay.ConflictChecker.TransactionRebaseState;
 import io.delta.kernel.internal.rowtracking.RowTracking;
+import io.delta.kernel.internal.tablefeatures.TableFeatures;
 import io.delta.kernel.internal.util.*;
 import io.delta.kernel.metrics.TransactionReport;
 import io.delta.kernel.types.StructType;
